[webgpu] s/ScatterOptimizedProgram/ScatterProgram/g
